32/64 bit handling

+++ Sources/ReaderContentPage.m
@@ -26,6 +26,7 @@
 #import "ReaderContentPage.h"
 #import "ReaderContentTile.h"
 #import "CGPDFDocument.h"
+#import <tgmath.h>
 
 
 @interface ReaderContentPage ()
@@ -244,10 +245,10 @@
 			}
 		}
 		
-		CGFloat vr_x = roundf(ll_x);
-		CGFloat vr_w = roundf(ur_x - ll_x);
-		CGFloat vr_y = roundf(ll_y);
-		CGFloat vr_h = roundf(ur_y - ll_y);
+		CGFloat vr_x = (CGFloat)round(ll_x);
+		CGFloat vr_w = (CGFloat)round(ur_x - ll_x);
+		CGFloat vr_y = (CGFloat)round(ll_y);
+		CGFloat vr_h = (CGFloat)round(ur_y - ll_y);
 		
 		CGRect viewRect = CGRectMake(vr_x, vr_y, vr_w, vr_h); // View CGRect from PDFRect
